Create README.md
This repository contains the Greek Integrated Modeling System (GRIMS), used in the study Koutsandreas et al. (2025): "A Multicriteria Modeling Framework for Evaluating Clean Energy Transitions: The Case of Greece as an Electricity Exporter." 

GRIMS is an integrated assessment model that represents Greece's power sector and its key interactions with other resource systems.

Additionally, the repository includes the key results from the examined scenarios, as well as the outcomes of the multicriteria framework applied in this study, which is based on the TOPSIS method.